requirements.txt
asgiref==3.4.0
Django==3.2.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
